feat: IC-56 - implemented macro for building packets; global refactoring (#30)
* IC-56 - implemented macro for building packets; global refactoring

* IC-56 - fixed issue with some packets; fixed config sample (were missed some required options there); refactored srp

* IC-56 - fixed crash on world server; support dynamic fields option (in progress); refactored code; refactored compressed option in macros

* IC-56 - added support for dynamic fields and dynamic opcodes; removed packet! macro in favor of proc macro usage; added json formatter and prettify output in debug details panel

* IC-56 - refactored types (including types for fields)

* IC-56 - improved formatting; fixed details output on scrolling history; refactoring

* IC-56 - improved formatting for update packets (show update field names); refactored terminated string

* IC-56 - refactored update fields to output float value properly

* IC-56 - added out_of_range/nearby objects guids to parsed block, improved output for packed_guid and some packets

* IC-56 - added fast scroll by PageUp/PageDown

* IC-56 - added scrolling for details panel (using ctrl + arrows)

* IC-56 - removed redundant rust-clippy check

* IC-56 - added support for home/end (to select first/last item in output); added info panel with extra details output; fixed freeze on switch between UI modes

+++ src/crypto/decryptor.rs
@@ -8,8 +8,8 @@ const DECRYPTION_KEY: [u8; 16] = [
     0xCC, 0x98, 0xAE, 0x04, 0xE8, 0x97, 0xEA, 0xCA, 0x12, 0xDD, 0xC0, 0x93, 0x42, 0x91, 0x53, 0x57
 ];
 
-pub const INCOMING_HEADER_LENGTH: u8 = 4;
-pub const INCOMING_OPCODE_LENGTH: u16 = 2;
+pub const INCOMING_HEADER_LENGTH: usize = 4;
+pub const INCOMING_OPCODE_LENGTH: usize = 2;
 
 pub struct Decryptor {
     instance: RC4,
@@ -31,8 +31,8 @@ impl Decryptor {
     }
 
     pub fn decrypt(&mut self, data: &[u8]) -> Vec<u8> {
-        let header = self.instance.encrypt(&data[..(INCOMING_HEADER_LENGTH as usize)]);
-        [header, data[(INCOMING_HEADER_LENGTH as usize)..].to_vec()].concat().to_vec()
+        let header = self.instance.encrypt(&data[..INCOMING_HEADER_LENGTH]);
+        [header, data[INCOMING_HEADER_LENGTH..].to_vec()].concat().to_vec()
     }
 }
